--- a/systems/memory-hierarchy-2/metrics/go.mod
+++ b/systems/memory-hierarchy-2/metrics/go.mod
@@ -1,0 +1,3 @@
+module metrics
+
+go 1.16