--- a/systems/memory-hierarchy-2/metrics/metrics_test.go
+++ b/systems/memory-hierarchy-2/metrics/metrics_test.go
@@ -6,12 +6,12 @@ import (
 )
 
 func BenchmarkMetrics(b *testing.B) {
-	users := LoadData()
+	users, userData := LoadData()
 
 	b.Run("Average age", func(b *testing.B) {
 		actual := 0.0
 		for n := 0; n < b.N; n++ {
-			actual = AverageAge(users)
+			actual = AverageAge(&userData)
 		}
 		expected := 59.62
 		if math.IsNaN(actual) || math.Abs(actual-expected) > 0.01 {