--- a/systems/memory-hierarchy-2/metrics/metrics_test.go
+++ b/systems/memory-hierarchy-2/metrics/metrics_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func BenchmarkMetrics(b *testing.B) {
-	users, userData := LoadData()
+	_, userData := LoadData()
 
 	b.Run("Average age", func(b *testing.B) {
 		actual := 0.0
@@ -22,7 +22,7 @@ func BenchmarkMetrics(b *testing.B) {
 	b.Run("Average payment", func(b *testing.B) {
 		actual := 0.0
 		for n := 0; n < b.N; n++ {
-			actual = AveragePaymentAmount(users)
+			actual = AveragePaymentAmount(&userData)
 		}
 
 		expected := 499850.559
@@ -34,7 +34,7 @@ func BenchmarkMetrics(b *testing.B) {
 	b.Run("Payment stddev", func(b *testing.B) {
 		actual := 0.0
 		for n := 0; n < b.N; n++ {
-			actual = StdDevPaymentAmount(users)
+			actual = StdDevPaymentAmount(&userData)
 		}
 		expected := 288684.850
 		if math.IsNaN(actual) || math.Abs(actual-expected) > 0.01 {