My results so far.

| Function | Original (ns/op) | Optimized (ns/op) | Speed-up
| --- | ---: | ---: | ---: |
| BenchmarkMetrics/Average_age-8            | 1,141,434        |  37,691    | 30.3x
| BenchmarkMetrics/Average_payment-8    | 17,101,507      | 292,921   | 58.4x
| BenchmarkMetrics/Payment_stddev-8      |  34,107,580   | 926,618   | 36.8x

Discussion of methodology in the README.